requirements.txt
openai
langchain
django
djangorestframework
chromadb
langchain_community
langchain-openai
requests
deep-translator
pytest

